+++ sources/packages/backend/apps/db-migrations/src/sql/Notifications/Add-metadata-column.sql
@@ -3,4 +3,4 @@ ALTER TABLE
 ADD
   COLUMN metadata jsonb;
 
-COMMENT ON COLUMN sims.notifications.metadata IS 'Metadata information on notifications sent such as tracking the various relations associated with the sent notification.';
+COMMENT ON COLUMN sims.notifications.metadata IS 'Metadata information for a notification to allow a more granular identification or identify further the reason why it was generated.';
