+++ sources/packages/backend/libs/integrations/src/services/disbursement-schedule/e-cert-calculation/e-cert-calculation-process.ts
@@ -1,20 +1,9 @@
 import { ECertProcessStep } from "../e-cert-processing-steps/e-cert-steps-models";
-import { DataSource, EntityManager } from "typeorm";
+import { DataSource } from "typeorm";
 import { ProcessSummary } from "@sims/utilities/logger";
-import {
-  addDays,
-  getUTC,
-  getUTCNow,
-  parseJSONError,
-  processInParallel,
-} from "@sims/utilities";
+import { parseJSONError, processInParallel } from "@sims/utilities";
 import { EligibleECertDisbursement } from "../disbursement-schedule.models";
-import { Notification, Student } from "@sims/sims-db";
-import {
-  DisbursementBlockedNotificationForMinistry,
-  NotificationActionsService,
-  StudentNotification,
-} from "@sims/services";
+import { ECertNotificationService } from "@sims/integrations/services";
 
 /**
  * Disbursements grouped by student to allow parallel processing of students
@@ -24,11 +13,6 @@ interface GroupedDisbursementPerStudent {
   [studentId: number]: EligibleECertDisbursement[];
 }
 
-interface NotificationData {
-  max_created_at: Date;
-  notificationsCount: number;
-}
-
 /**
  * Base process for e-Cert calculations independently of the offering intensity.
  * Both offering intensity will have similar calculation steps and some additional ones.
@@ -36,7 +20,7 @@ interface NotificationData {
 export abstract class ECertCalculationProcess {
   constructor(
     private readonly dataSource: DataSource,
-    private readonly notificationActionsService: NotificationActionsService,
+    private readonly ecertNotificationService: ECertNotificationService,
   ) {}
 
   /**
@@ -139,19 +123,12 @@ export abstract class ECertCalculationProcess {
               disbursementLog,
             );
             if (!shouldProceed) {
-              const shouldCreateNotification =
-                await this.shouldCreateDisbursementBlockedNotification(
-                  eCertDisbursement.disbursement.id,
-                  entityManager,
-                );
-              if (shouldCreateNotification) {
-                await this.createDisbursementBlockedNotifications(
-                  eCertDisbursement.disbursement.id,
-                  eCertDisbursement.studentId,
-                  eCertDisbursement.applicationNumber,
-                  entityManager,
-                );
-              }
+              await this.ecertNotificationService.createDisbursementBlockedNotification(
+                eCertDisbursement.disbursement.id,
+                eCertDisbursement.studentId,
+                eCertDisbursement.applicationNumber,
+                entityManager,
+              );
               disbursementLog.info(
                 "The step determined that the calculation should be interrupted. This disbursement will not be part of the next e-Cert generation.",
               );
@@ -168,99 +145,5 @@ export abstract class ECertCalculationProcess {
         break;
       }
     }
-  }
-
-  /**
-   * Checks whether the disbursement blocked notification should be created or not.
-   * @param disbursementId disbursement id.
-   * @param entityManager entity manager to execute in transaction.
-   * @returns boolean indicating if the disbursement blocked notification should be created or not.
-   */
-  private async shouldCreateDisbursementBlockedNotification(
-    disbursementId: number,
-    entityManager: EntityManager,
-  ): Promise<boolean> {
-    // Create notifications to the student and ministry when the disbursement is blocked.
-    const notificationMessageId = 16;
-    const notification = await entityManager
-      .getRepository(Notification)
-      .createQueryBuilder("notifications")
-      .select("count(*)::int", "notificationsCount")
-      .addSelect("max(notifications.created_at)", "max_created_at")
-      .innerJoin("notifications.notificationMessage", "notificationMessage")
-      .where("notificationMessage.id = :notificationMessageId", {
-        notificationMessageId,
-      })
-      .andWhere("notifications.permanent_failure_error IS NULL")
-      .andWhere(
-        "notifications.metadata->>'disbursement_id' = :disbursementId",
-        {
-          disbursementId,
-        },
-      )
-      .getRawOne<NotificationData>();
-    const sevenDaysLaterDate = getUTC(addDays(7, notification.max_created_at));
-    // Condition check to create notifications: Less than 3 notifications are created previously and there are no failures in sending those created notifications.
-    // Plus, it has been atleast 7 days from the last created notification for this disbursement.
-    if (
-      !notification.notificationsCount ||
-      (notification.notificationsCount < 3 && sevenDaysLaterDate < getUTCNow())
-    ) {
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * Creates disbursement blocked notifications for student and ministry.
-   * @param disbursementId disbursement id associated with the blocked disbursement.
-   * @param studentId student id associated with the blocked disbursement.
-   * @param applicationNumber application number associated with the blocked disbursement.
-   * @param entityManager entity manager to execute in transaction.
-   */
-  private async createDisbursementBlockedNotifications(
-    disbursementId: number,
-    studentId: number,
-    applicationNumber: string,
-    entityManager: EntityManager,
-  ) {
-    const student = await entityManager.getRepository(Student).findOne({
-      select: {
-        id: true,
-        birthDate: true,
-        user: { firstName: true, lastName: true, email: true },
-      },
-      relations: { user: true },
-      where: { id: studentId },
-    });
-    const studentNotification: StudentNotification = {
-      givenNames: student.user.firstName,
-      lastName: student.user.lastName,
-      toAddress: student.user.email,
-      userId: student.user.id,
-    };
-    const ministryNotification: DisbursementBlockedNotificationForMinistry = {
-      givenNames: student.user.firstName,
-      lastName: student.user.lastName,
-      email: student.user.email,
-      dob: student.birthDate,
-      applicationNumber: applicationNumber,
-    };
-    const disbursementBlockedNotificationForStudentPromise =
-      this.notificationActionsService.saveDisbursementBlockedNotificationForStudent(
-        studentNotification,
-        disbursementId,
-        entityManager,
-      );
-    const disbursementBlockedNotificationForMinistryPromise =
-      this.notificationActionsService.saveDisbursementBlockedNotificationForMinistry(
-        ministryNotification,
-        disbursementId,
-        entityManager,
-      );
-    await Promise.all([
-      disbursementBlockedNotificationForStudentPromise,
-      disbursementBlockedNotificationForMinistryPromise,
-    ]);
   }
 }
