+++ sources/packages/backend/libs/integrations/src/services/disbursement-schedule/e-cert-calculation/full-time-calculation-process.ts
@@ -16,6 +16,7 @@ import { ECertGenerationService } from "../e-cert-generation.service";
 import { Injectable } from "@nestjs/common";
 import { EligibleECertDisbursement } from "../disbursement-schedule.models";
 import { NotificationActionsService } from "@sims/services";
+import { ECertNotificationService } from "../e-cert-notification/e-cert-notification.service";
 
 /**
  * Executes the full-time calculations for the e-Cert generation.
@@ -24,7 +25,7 @@ import { NotificationActionsService } from "@sims/services";
 export class FullTimeCalculationProcess extends ECertCalculationProcess {
   constructor(
     dataSource: DataSource,
-    notificationActionsService: NotificationActionsService,
+    ecertNotificationService: ECertNotificationService,
     private readonly eCertGenerationService: ECertGenerationService,
     private readonly validateDisbursementFullTimeStep: ValidateDisbursementFullTimeStep,
     private readonly applyOverawardsDeductionsStep: ApplyOverawardsDeductionsStep,
@@ -35,7 +36,7 @@ export class FullTimeCalculationProcess extends ECertCalculationProcess {
     private readonly createBCTotalGrantsStep: CreateBCTotalGrantsStep,
     private readonly persistCalculationsStep: PersistCalculationsStep,
   ) {
-    super(dataSource, notificationActionsService);
+    super(dataSource, ecertNotificationService);
   }
 
   /**
