+++ sources/packages/backend/libs/integrations/src/services/index.ts
@@ -16,3 +16,4 @@ export * from "./atbc/atbc.service";
 export * from "./atbc/models/atbc.model";
 export * from "./sin-validation/sin-validation.service";
 export * from "./institution-location/institution-location.service";
+export * from "./disbursement-schedule/e-cert-notification/e-cert-notification.service";
