+++ sources/packages/backend/libs/services/src/constants/index.ts
@@ -3,3 +3,4 @@ export * from "./error-code.constants";
 export * from "./disbursements.constants";
 export * from "./worker-constants";
 export * from "./restriction.constants";
+export * from "./notification.constants";
