+++ sources/packages/backend/libs/services/src/constants/notification.constants.ts
@@ -1,0 +1,2 @@
+// Represents the minimum number of days before another notification for a blocked disbursement will be sent out.
+export const BLOCKED_DISBURSEMENT_NOTIFICATION_MIN_DAYS_INTERVAL = 7;
